project1/requirements.txt
Django==2.0.1
pip==9.0.1
pytz==2017.3
setuptools==38.4.0
wheel==0.30.0
python-twitter==3.3
praw==5.3.0
requests
wikipedia-api==0.3.4
imdbpie==5.2.0
